requirement.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pwdlib[argon2]>=0.3.0
PyJWT>=2.8.0
